XPRT - Simplify refcounting
The epoll loop needs a ref on the XPRT, so that it stays alive while in
epoll().  Previously, the ref was taken and dropped across individual
events.  Instead, take a ref when hooking, and keep it until it's
unhooked.  This means up to 2 refs for an XPRT (one for send and one for
receive).  This simplifies refcounting, making it symmetric everywhere,
rather than passing the refcount for an event off to the processor to be
released.

In addition, there is a race when unhooking events from epoll, where the
event could be ready for delivery (or even delivered, but the thread not
scheduled) and so the event is processed after the unhook, and therefore
after the XPRT has been freed.  To close this, stop putting a pointer to
the rec in the event data, and instead put the FD in there and use it to
look up the XPRT.  This ensures that, if we got the XPRT from lookup,
it's valid and ref'd for the duration of the event.

Signed-off-by: Daniel Gryniewicz <dang@redhat.com>
(cherry picked from commit 57aa974bc1799d794159b07e3f9c56bc2a835af2)

[ Since we use fd for the event instead of xprt/rec, we don't need a ref
in hook, unhook or rearm functions. ]
Signed-off-by: Malahal Naineni <mnaineni@in.ibm.com>